Refactor and improve widening type problem reporting
Centralizes and refines helper functions for detecting and
reporting serialization problems related to widening codecs.
This improves code organization and enables more precise
diagnostic messages, particularly for `MapProperty` keys and values,
by allowing custom resolution suggestions.
