+++ platforms/core-configuration/core-serialization-codecs/src/main/kotlin/org/gradle/internal/serialize/codecs/core/ProviderCodecs.kt
@@ -52,7 +52,6 @@ import org.gradle.internal.build.BuildStateRegistry
 import org.gradle.internal.cc.base.serialize.IsolateOwners
 import org.gradle.internal.configuration.problems.PropertyTrace
 import org.gradle.internal.extensions.core.serviceOf
-import org.gradle.internal.reflect.UnsupportedTypeException
 import org.gradle.internal.extensions.stdlib.uncheckedCast
 import org.gradle.internal.file.PathToFileResolver
 import org.gradle.internal.flow.services.BuildWorkResultProvider
@@ -62,11 +61,9 @@ import org.gradle.internal.serialize.graph.IsolateContext
 import org.gradle.internal.serialize.graph.MutableIsolateContext
 import org.gradle.internal.serialize.graph.ReadContext
 import org.gradle.internal.serialize.graph.WriteContext
-import org.gradle.internal.serialize.graph.codecs.findCodecThatWidensIncompatibly
-import org.gradle.internal.serialize.graph.taskDescription
+import org.gradle.internal.serialize.graph.reportIfUnsupportedPropertyValueType
 import org.gradle.internal.serialize.graph.codecs.BeanCodec
 import org.gradle.internal.serialize.graph.codecs.Bindings
-import org.gradle.internal.serialize.graph.codecs.WideningCodec
 import org.gradle.internal.serialize.graph.decodeBean
 import org.gradle.internal.serialize.graph.decodePreservingIdentity
 import org.gradle.internal.serialize.graph.decodePreservingSharedIdentity
@@ -74,7 +71,6 @@ import org.gradle.internal.serialize.graph.encodeBean
 import org.gradle.internal.serialize.graph.encodePreservingIdentityOf
 import org.gradle.internal.serialize.graph.encodePreservingSharedIdentityOf
 import org.gradle.internal.serialize.graph.logPropertyProblem
-import org.gradle.internal.serialize.graph.reportSerializationProblem
 import org.gradle.internal.serialize.graph.readClassOf
 import org.gradle.internal.serialize.graph.readNonNull
 import org.gradle.internal.serialize.graph.runReadOperation
@@ -408,33 +404,6 @@ abstract class AbstractPropertyCodec<P : AbstractProperty<*, *>>(
 }
 
 
-/**
- * Reports a deferred problem when the codec registered for [valueType] is a
- * [WideningCodec] that produces a decoded type not assignable to [valueType].
- *
- * @return `true` when the property's value must be dropped from the cache
- *         (the caller should write a missing-value placeholder so the
- *         property survives the roundtrip as if it were never set).
- */
-internal suspend fun WriteContext.reportIfUnsupportedPropertyValueType(propertyKind: Class<*>, valueType: Class<*>): Boolean {
-    val widening = findCodecThatWidensIncompatibly(valueType) ?: return false
-
-    val resolution = if (MapProperty::class.java.isAssignableFrom(propertyKind)) {
-        "Avoid using ${valueType.simpleName} as a MapProperty key or value."
-    } else {
-        widening.wideningFix
-    }
-    val exception = UnsupportedTypeException(
-        "Cannot serialize ${propertyKind.simpleName}<${valueType.simpleName}> in ${trace.taskDescription()}. " +
-            "The value type of this property (${valueType.name}) is not supported with the configuration cache: " +
-            "its codec produces ${widening.publicDecodedType.name} on load.",
-        listOf(resolution)
-    )
-    reportSerializationProblem(exception)
-    return true
-}
-
-
 class PropertyCodec(
     private val propertyFactory: PropertyFactory,
     providerCodec: FixedValueReplacingProviderCodec
@@ -594,8 +563,12 @@ class MapPropertyCodec(
 
     override suspend fun WriteContext.encodeThis(value: DefaultMapProperty<*, *>) {
         encodePreservingIdentityOf(value) {
-            val keyUnsupported = reportIfUnsupportedPropertyValueType(MapProperty::class.java, value.keyType)
-            val valueUnsupported = reportIfUnsupportedPropertyValueType(MapProperty::class.java, value.valueType)
+            val keyUnsupported = reportIfUnsupportedPropertyValueType(MapProperty::class.java, value.keyType) { _, v ->
+                mapPropertyResolutionFor("key", v)
+            }
+            val valueUnsupported = reportIfUnsupportedPropertyValueType(MapProperty::class.java, value.valueType) { _, v ->
+                mapPropertyResolutionFor("value", v)
+            }
             writeClass(value.keyType)
             writeClass(value.valueType)
             providerCodec.run {
@@ -604,6 +577,12 @@ class MapPropertyCodec(
             }
         }
     }
+
+    // MapProperty's user-facing fix names BOTH the offending type and whether the
+    // problem is with the map's key or the map's value, so users can act precisely
+    // — overriding the codec's bean-field-oriented `wideningFix`.
+    private fun mapPropertyResolutionFor(kind: String, valueType: Class<*>): String =
+        "Avoid using ${valueType.simpleName} as a MapProperty $kind."
 
     override suspend fun ReadContext.decodeThis(): DefaultMapProperty<*, *> {
         return decodePreservingIdentity { id ->
